requirements.txt
future
pillow
numpy<2.0,>=1.14.5
setuptools>=41.0.0
torch==1.2
tensorflow>=1.14
torchtext
sacrebleu>=1.3.6
subword-nmt
matplotlib
seaborn
pyyaml>=5.1
pylint
entmax>=1.0
